+++ Mikibot.Crawler/WebsocketCrawler/Packet/BasePacket.cs
@@ -46,7 +46,7 @@ namespace Mikibot.Crawler.WebsocketCrawler.Packet
             Version = 0,
         };
 
-        public static BasePacket Auth(int roomId, string auth)
+        public static BasePacket Auth(long roomId, string auth)
         {
             var authPacket = JsonSerializer.Serialize(new
             {
